+++ HISTORY.rst
@@ -4,6 +4,8 @@ HISTORY
 0.0.x (2015-xx-xx)
 ------------------
 * Added method to class description.
+* If SeqRecord does not have the `reading_frame` parameter, it will issue a warning.
+  It used to issue an exception.
 
 0.0.2 (2015-09-13)
 ------------------